+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/world/seasons/Seasons.java
@@ -81,6 +81,10 @@ public class Seasons extends Feature {
 	public static Season.SubSeason startingSeason = Season.SubSeason.EARLY_SUMMER;
 
 	@Config(min = 0, max = 3d)
+	@Label(name = "Time Control day night duration", description = "How many minutes will day and night duration be (with this set to 10 the day will last 10 minutes and the night 10 minutes for a grand total of 20 minutes). This also adjusts the day_duration config option in Serene Seasons. Vanilla is 10.")
+	public static Double timeControlDayNightDuration = 10d;
+
+	@Config(min = 0, max = 3d)
 	@Label(name = "Time Control day night shift", description = "How many minutes will day and night duration be shifted based off seasons? E.g. in Mid spring / autumn the duration of day and night is vanilla, when moving off those seasons day and night will last this many minutes more/less. In mid summer / winter the duration of day and night duration will be more / less by 3 times this value. Set to 0 to disable")
 	public static Double timeControlDayNightShift = 1d;
 
@@ -104,6 +108,8 @@ public class Seasons extends Feature {
 
 		if (changeSereneSeasonsConfig)
 			FertilityConfig.seasonalCrops.set(false);
+		if (ModList.get().isLoaded("timecontrol") && timeControlDayNightDuration != 10)
+			ServerConfig.dayDuration.set((int) (timeControlDayNightDuration * 2d));
 	}
 
 	@SubscribeEvent