+++ changelog.md
@@ -6,6 +6,8 @@
   * Renamed some advancements
   * Moved some others
   * Removed Ore Rocks advancement
+* Added 'Time Control day night duration'
+  * Allows to set the duration of day with Time Control
 * Removed `iguanatweaksexpanded:iron_ores`, `gold_ores` and `copper_ores` block tags. These tags already exist in minecraft
 * Added tooltip to the firestarter
 * Fixed cleansed lapis drop with Fortune I being too high